+++ tests/extractors/test_python.py
@@ -153,6 +153,16 @@ def test_tagged_multiline_comment():
 
 
 @pytest.mark.usefixtures('fake_source')
+def test_flags_in_comment():
+    global source
+    options = mock.Mock()
+    options.comment_tag = 'I18N:'
+    source = u'''# I18N: [markdown-format,fuzzy] Comment]\n_(u'word')'''
+    messages = list(python_extractor('filename', options))
+    assert messages[0].flags == ['markdown-format', 'fuzzy']
+
+
+@pytest.mark.usefixtures('fake_source')
 def test_comment_and_default_value():
     global source
     options = mock.Mock()